d63e31c refactor(@angular/ssr): remove deprecated `provideServerRoutesConfig` dev-preview API

+++ migrations/use-application-builder/migration.js
@@ -99,7 +99,10 @@ function* updateBuildTarget(projectName, buildTarget, serverTarget, tree, contex
     // Update server file
     const ssrMainFile = serverTarget?.options?.['main'];
     if (typeof ssrMainFile === 'string') {
-        yield deleteFile(ssrMainFile);
+        // Do not delete the server main file if it's the same as the browser file.
+        if (buildTarget.options?.browser !== ssrMainFile) {
+            yield deleteFile(ssrMainFile);
+        }
         yield (0, schematics_1.externalSchematic)('@schematics/angular', 'ssr', {
             project: projectName,
             skipInstall: true,
